+++ src/imio/dashboard/events.py
@@ -13,3 +13,8 @@ from imio.helpers.cache import invalidate_cachekey_volatile_for
 def onDashboardCollectionModified(obj, event):
     '''Called whenever a DashboardCollection is modified.'''
     invalidate_cachekey_volatile_for('imio.dashboard.conditionawarecollectionvocabulary')
+
+
+def onDashboardCollectionTransition(obj, event):
+    '''Called whenever a WF transition was triggered on a DashboardCollection.'''
+    invalidate_cachekey_volatile_for('imio.dashboard.conditionawarecollectionvocabulary')